Fetch upstream branch and fixed CI tests
