+++ src/client/app/components/LanguageSelectorComponent.tsx
@@ -8,7 +8,6 @@ import { LanguageTypes } from '../types/redux/i18n';
 import { UpdateDefaultLanguageAction } from '../types/redux/admin';
 import { FormattedMessage } from 'react-intl';
 import TooltipMarkerComponent from './TooltipMarkerComponent';
-import TooltipHelpContainerAlternative from '../containers/TooltipHelpContainerAlternative';
 import Dropdown from 'reactstrap/lib/Dropdown';
 import DropdownItem from 'reactstrap/lib/DropdownItem';
 import DropdownToggle from 'reactstrap/lib/DropdownToggle';
@@ -57,7 +56,6 @@ export default class LanguageSelectorComponent extends React.Component<LanguageS
 
 		return (
 			<div style={divBottomPadding}>
-				<TooltipHelpContainerAlternative page='home' />
 				<p style={labelStyle}>
 					<FormattedMessage id='language' />:
 				</p>
